refactor(unauthed/user): Move endpoints to mod.rs Redundant to have it in a separate file imo
